🚚 rename: WeeklyMatches 컴포넌트 파일 이름 index로 변경 #40
